basic ui components and logic added
basic ui components doesn't account for some specific images below Scully, as well as a fleshed out Scully textbox.

However, the dividers and boards for P1 and P2 are ready and the logic for warping the UI to be what we want is also ready.

Fleshing out needed but it's a good 70% of the way there if not more